Mechanical change: move ygot/proto -> protomap.
 * (M) ygot/proto(_test)?.go
  - Move protobuf mapping code into its own package.
